+++ app/src/main/res/layout/fragment_daily_weather.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.next_days_weather.DailyFragment">
+    tools:context=".ui.daily_weather.DailyFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
